refactor: clean up code
- "merges" slash and interaction contexts into super class
- fixes a bug in timestamp logic of ccv3
